+++ src/datamasque_cli/main.py
@@ -19,6 +19,7 @@ from datamasque_cli.commands import (
     connections,
     discovery,
     files,
+    ifm,
     ruleset_libraries,
     rulesets,
     runs,
@@ -43,6 +44,7 @@ app.add_typer(seeds.app, name="seeds")
 app.add_typer(files.app, name="files")
 app.add_typer(system.app, name="system")
 app.add_typer(ruleset_libraries.app, name="libraries")
+app.add_typer(ifm.app, name="ifm")
 
 
 @app.command()